// src/app/models/ownership-chain.model.ts
export class OwnershipchainModel {
  ExecutantName?: string;
  DateOfRegistration?: string;
  DocumentNumber?: string;
  TransactionNature?: string;
  ChainPropertySize?: string;
  SurveyNumbers?: string;
  ChainMarketValue?: string;
  ChainConsidrationValue?: string;
  ClaimantName?: string;
  Children?: Array<OwnershipchainModel>;
}